fix: use int64 for release id in openapi spec
Mark the releaseId path parameter as int64 for getReleaseById and regenerate SDK/server artifacts for v0.3.2.

Also guard the Spring Boot delegate compatibility fallback so out-of-range int64 release ids return a controlled 400 instead of surfacing an ArithmeticException.

+++ spring-boot2/src/main/java/com/apollo/openapi/server/api/ReleaseManagementApi.java
@@ -294,7 +294,7 @@ public interface ReleaseManagementApi {
     )
     default ResponseEntity<OpenReleaseDTO> getReleaseById(
         @Parameter(name = "env", description = "环境标识", required = true, in = ParameterIn.PATH) @PathVariable("env") String env,
-        @Parameter(name = "releaseId", description = "发布ID", required = true, in = ParameterIn.PATH) @PathVariable("releaseId") Integer releaseId
+        @Parameter(name = "releaseId", description = "发布ID", required = true, in = ParameterIn.PATH) @PathVariable("releaseId") Long releaseId
     ) {
         return getDelegate().getReleaseById(env, releaseId);
     }
